Change vars to floats or doubles, found maps that violate the .osu spec

+++ OsuDiffCalc/FileProcessor/Analyzer.cs
@@ -307,7 +307,7 @@
 			return difficulty;
 		}
 
-		static float GetJumpDifficulty(float distance, float time, Beatmap map) {
+		static float GetJumpDifficulty(double distance, double time, Beatmap map) {
 			if (time <= 0)
 				return 0;
