+++ OsuDiffCalc/FileProcessor/BeatmapObjects/HitObject.cs
@@ -7,7 +7,7 @@
 	/// </summary>
 	/// <param name="X"> X initial position of the HitObject in osupixels </param>
 	/// <param name="Y"> Y initial position of the HitObject in osupixels </param>
-	abstract record HitObject(int X, int Y, int StartTime, int EndTime) : BeatmapObject(StartTime, EndTime) {
+	abstract record HitObject(float X, float Y, double StartTime, double EndTime) : BeatmapObject(StartTime, EndTime) {
 		public override string ToString() => $"{GetType().Name}:  xy({X} {Y})  time({TimingParser.GetTimeStamp(StartTime)} {TimingParser.GetTimeStamp(EndTime)})";
 	}
 }
